# Conflicts:
#	src/main/java/com/itwillbs/ilkwangtech/sales/dto/CustomerDTO.java
#	src/main/java/com/itwillbs/ilkwangtech/sales/repository/CustomerRepository.java
#	src/main/java/com/itwillbs/ilkwangtech/sales/service/CustomerService.java
#	src/main/java/com/itwillbs/ilkwangtech/sales/service/CustomerServiceImpl.java

+++ src/main/java/com/itwillbs/ilkwangtech/account/repository/CommonCodeRepository.java
@@ -3,7 +3,7 @@ package com.itwillbs.ilkwangtech.account.repository;
 import com.itwillbs.ilkwangtech.common.entity.CommonCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommonCodeRepository extends JpaRepository<CommonCode, Integer> {
+public interface CommonCodeRepository extends JpaRepository<CommonCode, Long> {
